moves all packages from hatdex to dataswift namespace
