+++ src/main/scala/io/dataswift/dex/api/services/DexClient.scala
@@ -7,25 +7,21 @@
  *
  */
 
-package org.hatdex.dex.apiV2.services
+package io.dataswift.dex.api.services
 
 import javax.inject.Inject
 
-import org.hatdex.dex.api.services.{ DexDataPlugs, DexNotices }
 import play.api.libs.ws.WSClient
 
 class DexClient(
     val ws: WSClient,
     val dexAddress: String,
-    override val schema: String,
-    override val apiVersion: String) extends DexOffers
+    override val schema: String) extends DexOffers
   with DexNotices
   with DexDataPlugs
-  with DexStats
-  with DexUsers
-  with DexApplications {
+  with DexStats {
 
-  @Inject def this(ws: WSClient, dexAddress: String) = this(ws, dexAddress, "https://", "v1.1")
+  @Inject def this(ws: WSClient, dexAddress: String) = this(ws, dexAddress, "https://")
 
   val logger = play.api.Logger(this.getClass)
 }
