+++ src/main/scala/io/dataswift/dex/api/services/DexNotices.scala
@@ -7,10 +7,9 @@
  *
  */
 
-package org.hatdex.dex.api.services
+package io.dataswift.dex.api.services
 
-import org.hatdex.dex.api.json.DexJsonFormats
-import org.hatdex.dex.api.models.Notice
+import io.dataswift.dex.api.models.Notice
 import play.api.Logger
 import play.api.http.Status._
 import play.api.libs.json.Json
@@ -24,7 +23,7 @@ trait DexNotices {
   val schema: String
   val dexAddress: String
 
-  import DexJsonFormats.noticeFormat
+  import io.dataswift.dex.api.json.DexJsonFormats.noticeFormat
 
   def postNotice(access_token: String, notice: Notice)(implicit ec: ExecutionContext): Future[Notice] = {
     logger.debug(s"Post notice $notice to MarketSquare")
